refractor activities into packages
