+++ app/src/main/java/com/appr/digibiz/adapters/PageAdapter.java
@@ -1,9 +1,11 @@
-package com.appr.digibiz;
+package com.appr.digibiz.adapters;
 
-import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
+
+import com.appr.digibiz.fragments.AvailableFragment;
+import com.appr.digibiz.fragments.OutOfStockFragment;
 
 public class PageAdapter extends FragmentPagerAdapter {
     private int numOfTabs;
