+++ app/src/main/java/com/appr/digibiz/fragments/AvailableFragment.java
@@ -1,5 +1,6 @@
-package com.appr.digibiz;
+package com.appr.digibiz.fragments;
 
+import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -11,36 +12,36 @@ import android.view.ViewGroup;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.appr.digibiz.R;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class OutOfStockFragment extends Fragment implements View.OnClickListener{
-    @BindView(R.id.homeProgressBar) ProgressBar mHomeProgressBar;
-    @BindView(R.id.outOfStockListRecyclerView)
-    RecyclerView outOfStockListRecyclerView;
+public class AvailableFragment extends Fragment implements View.OnClickListener{
+    @BindView(R.id.homeProgressBar)
+    ProgressBar mHomeProgressBar;
+    @BindView(R.id.availableListRecyclerView)
+    RecyclerView mAvailableListRecyclerView;
     @BindView(R.id.homeErrorText)
     TextView mErrorText;
     @BindView(R.id.fab)
     FloatingActionButton mFab;
 
-
-    public OutOfStockFragment() {
+    public AvailableFragment() {
         // Required empty public constructor
     }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_out_of_stock, container, false);
+        View view = inflater.inflate(R.layout.fragment_available, container, false);
         ButterKnife.bind(this, view);
 
         //click listeners
