+++ app/src/main/java/com/appr/digibiz/fragments/OutOfStockFragment.java
@@ -1,6 +1,5 @@
-package com.appr.digibiz;
+package com.appr.digibiz.fragments;
 
-import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.fragment.app.Fragment;
@@ -12,35 +11,37 @@ import android.view.ViewGroup;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
+import com.appr.digibiz.R;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class AvailableFragment extends Fragment implements View.OnClickListener{
-    @BindView(R.id.homeProgressBar)
-    ProgressBar mHomeProgressBar;
-    @BindView(R.id.availableListRecyclerView)
-    RecyclerView mAvailableListRecyclerView;
+public class OutOfStockFragment extends Fragment implements View.OnClickListener{
+    @BindView(R.id.homeProgressBar) ProgressBar mHomeProgressBar;
+    @BindView(R.id.outOfStockListRecyclerView)
+    RecyclerView outOfStockListRecyclerView;
     @BindView(R.id.homeErrorText)
     TextView mErrorText;
     @BindView(R.id.fab)
     FloatingActionButton mFab;
 
-    public AvailableFragment() {
+
+    public OutOfStockFragment() {
         // Required empty public constructor
     }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_available, container, false);
+        View view = inflater.inflate(R.layout.fragment_out_of_stock, container, false);
         ButterKnife.bind(this, view);
 
         //click listeners
